+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* fixed issue in task version update with optional arguments (\#475).
+
 # 1.0.1
 
 * used payload containing all fields in meta properties PATCH endpoint (\#473).